+++ middleware/InterfacePlayerRDK.h
@@ -146,7 +146,7 @@ enum class InterfaceCB
 	startNewSubtitleStream // Add more events here if needed
 };
 
-
+// Codec information structure
 struct CodecInfo
 {
 	GstStreamOutputFormat codecFormat; // 'avc1', 'mp4a', etc
@@ -166,20 +166,31 @@ struct CodecInfo
 			uint16_t height;
 		} video;
 	} info;
+
+	CodecInfo() : codecFormat(GST_FORMAT_INVALID), codecData(), isEncrypted(false)
+	{
+		memset(&info, 0, sizeof(info));
+	}
 };
 
+// DRM metadata structure
 struct MediaDrmMetadata
 {
 	bool isEncrypted;
 	std::string keyId; // 16 bytes UUID
 	std::vector<uint8_t> iv; // 8 or 16 bytes
 	std::string cipher; // e.g. 'cenc', 'cbcs'
-	std::string originalMediaType; // 'vide' or 'soun'
 	std::vector<uint8_t> subSamples; // optional subsample encryption data
 	uint8_t cryptByteBlock;
 	uint8_t skipByteBlock;
+
+	MediaDrmMetadata() : isEncrypted(false), keyId(), iv(), cipher(),
+		subSamples(), cryptByteBlock(0), skipByteBlock(0)
+	{
+	}
 };
 
+// Media sample structure
 struct MediaSample
 {
 	const void *data;
@@ -188,7 +199,11 @@ struct MediaSample
 	double dts;
 	double duration;
 	double ptsOffset;
-    MediaDrmMetadata drmMetadata; // DRM metadata
+	MediaDrmMetadata drmMetadata;
+
+	MediaSample() : data(nullptr), dataSize(0), pts(0.0), dts(0.0), duration(0.0), ptsOffset(0.0), drmMetadata()
+	{
+	}
 };
 
 // Class to encapsulate GStreamer-related functionality
@@ -804,6 +819,11 @@ class InterfacePlayerRDK
 		 */
 		const MonitorAVState& GetMonitorAVState();
 
+		/**
+		 * @brief Sets the stream capabilities.
+		 * @param[in] type The media type.
+		 * @param[in] codecInfo The codec information.
+		 */
 		void SetStreamCaps(GstMediaType type, const CodecInfo &codecInfo);
 
 	private:
